Moving PersonnelGenerator to its own file
