+++ MekHQ/src/mekhq/campaign/personnel/generator/AbstractPersonnelGenerator.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with MekHQ.  If not, see <http://www.gnu.org/licenses/>.
  */
-package mekhq.campaign.personnel;
+package mekhq.campaign.personnel.generator;
 
 import java.util.Calendar;
 import java.util.GregorianCalendar;
@@ -28,6 +28,8 @@ import megamek.common.Crew;
 import mekhq.Utilities;
 import mekhq.campaign.Campaign;
 import mekhq.campaign.RandomSkillPreferences;
+import mekhq.campaign.personnel.Person;
+import mekhq.campaign.personnel.SkillType;
 
 /**
  * Represents a class which can generate new {@link Person} objects
