+++ MekHQ/src/mekhq/campaign/personnel/generator/DefaultPersonnelGenerator.java
@@ -16,11 +16,13 @@
  * You should have received a copy of the GNU General Public License
  * along with MekHQ.  If not, see <http://www.gnu.org/licenses/>.
  */
-package mekhq.campaign.personnel;
+package mekhq.campaign.personnel.generator;
 
 import java.util.Objects;
 
 import mekhq.campaign.Campaign;
+import mekhq.campaign.personnel.*;
+import mekhq.campaign.personnel.generator.AbstractPersonnelGenerator;
 import mekhq.campaign.universe.AbstractFactionSelector;
 import mekhq.campaign.universe.AbstractPlanetSelector;
 import mekhq.campaign.universe.DefaultFactionSelector;
